Déplace des actions vers espèces.js
